frontend: Combine ApartmentSalePage and LoadedApartmentSalePage files
